+++ karpiu/planning/optim/__init__.py
@@ -1,0 +1,2 @@
+from .netprofit_maximizer import RevenueMaximizer, RevenueMaximizer2
+from .target_maximizer import TargetMaximizer
